+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - [CL-767] Add data from backend to representative dashboard
 
+## 2022-06-13
+
+### Changed
+
+- [TEC-11] Upgraded react-router frontend dependency
+
 ## 2022-06-08_2
 
 ### Fixed